refactor(lifecycle): align screen lifecycle structure with next
