+++ packages/react-native-screen-transitions/src/shared/components/screen-lifecycle/hooks/history/navigator-route-registry.ts
