+++ packages/react-native-screen-transitions/src/shared/components/screen-lifecycle/hooks/use-close-transition-intent.ts
@@ -1,0 +1,205 @@
+/** biome-ignore-all lint/correctness/useHookAtTopLevel: STACK_TYPE is stable per navigator */
+import { useLayoutEffect, useRef } from "react";
+import {
+	runOnJS,
+	useAnimatedReaction,
+	useDerivedValue,
+	useFrameCallback,
+} from "react-native-reanimated";
+import { useSharedValueState } from "../../../hooks/reanimated/use-shared-value-state";
+import useStableCallback from "../../../hooks/use-stable-callback";
+import { useGestureContext } from "../../../providers/gestures";
+import {
+	type BaseDescriptor,
+	useDescriptorDerivations,
+} from "../../../providers/screen/descriptors";
+import { useStackCoreContext } from "../../../providers/stack/core.provider";
+import { useManagedStackContext } from "../../../providers/stack/managed.provider";
+import type { AnimationStoreMap } from "../../../stores/animation.store";
+import type { SystemStoreMap } from "../../../stores/system.store";
+import { StackType } from "../../../types/stack.types";
+import { animateToProgress } from "../../../utils/animation/animate-to-progress";
+import { resetStoresForScreen } from "./helpers/reset-stores-for-screen";
+
+interface CloseHookParams {
+	current: BaseDescriptor;
+	animations: AnimationStoreMap;
+	system: SystemStoreMap;
+	resetStores: () => void;
+	activateHighRefreshRate: () => void;
+	deactivateHighRefreshRate: () => void;
+}
+
+/**
+ * Managed close - reacts to closingRouteKeysShared from ManagedStackContext.
+ * Used by blank-stack and component-stack.
+ */
+const useManagedClose = ({
+	current,
+	animations,
+	system,
+	resetStores,
+	activateHighRefreshRate,
+	deactivateHighRefreshRate,
+}: CloseHookParams) => {
+	const { handleCloseRoute, closingRouteKeysShared } = useManagedStackContext();
+	const routeKey = current.route.key;
+
+	const handleManagedCloseEnd = useStableCallback((finished: boolean) => {
+		if (!finished) return;
+		handleCloseRoute({ route: current.route });
+		requestAnimationFrame(() => {
+			deactivateHighRefreshRate();
+			resetStores();
+		});
+	});
+
+	useAnimatedReaction(
+		() => {
+			const keys = closingRouteKeysShared.value;
+			return keys?.includes(routeKey) ?? false;
+		},
+		(isClosing, wasClosing) => {
+			if (!isClosing || wasClosing) return;
+
+			runOnJS(activateHighRefreshRate)();
+			animateToProgress({
+				target: "close",
+				spec: current.options.transitionSpec,
+				animations,
+				targetProgress: system.targetProgress,
+				onAnimationFinish: handleManagedCloseEnd,
+			});
+		},
+	);
+
+	return { handleManagedCloseEnd };
+};
+
+/**
+ * Native stack close - listens to beforeRemove navigation event.
+ */
+const useNativeStackClose = ({
+	current,
+	animations,
+	system,
+	resetStores,
+	activateHighRefreshRate,
+	deactivateHighRefreshRate,
+}: CloseHookParams) => {
+	const gestureContext = useGestureContext();
+	const pendingActionRef = useRef<any>(null);
+
+	const isAncestorDismissingViaGesture = useSharedValueState(
+		useDerivedValue(() => {
+			"worklet";
+			return (
+				gestureContext?.ancestorContext?.gestureAnimationValues.dismissing
+					?.value ?? false
+			);
+		}),
+	);
+
+	const handleNativeCloseEnd = useStableCallback((finished: boolean) => {
+		deactivateHighRefreshRate();
+
+		if (!finished || !pendingActionRef.current) {
+			pendingActionRef.current = null;
+			return;
+		}
+
+		current.navigation.dispatch(pendingActionRef.current);
+		pendingActionRef.current = null;
+		requestAnimationFrame(() => {
+			resetStores();
+		});
+	});
+
+	const handleBeforeRemove = useStableCallback((e: any) => {
+		const options = current.options as { enableTransitions?: boolean };
+		const isEnabled = options.enableTransitions;
+		const navigation = current.navigation;
+		const routeKey = current.route.key;
+		const state = navigation.getState();
+		const routeIndex = state.routes.findIndex(
+			(route) => route.key === routeKey,
+		);
+		const isFirstScreen = routeIndex <= 0;
+
+		if (!isEnabled || isAncestorDismissingViaGesture || isFirstScreen) {
+			animations.closing.set(1);
+			resetStores();
+			return;
+		}
+
+		e.preventDefault();
+		pendingActionRef.current = e.data.action;
+		activateHighRefreshRate();
+
+		animateToProgress({
+			target: "close",
+			spec: current.options.transitionSpec,
+			animations,
+			targetProgress: system.targetProgress,
+			onAnimationFinish: handleNativeCloseEnd,
+		});
+	});
+
+	// biome-ignore lint/correctness/useExhaustiveDependencies: navigation listener should only rebind when the navigator instance changes
+	useLayoutEffect(() => {
+		return current.navigation.addListener?.("beforeRemove", handleBeforeRemove);
+	}, [current.navigation]);
+
+	return { handleNativeCloseEnd };
+};
+
+/**
+ * Handles close transition intent and returns finish callbacks for cleanup.
+ *
+ * Phase 1 keeps the existing v3 immediate-close behavior while moving the code
+ * behind the same hook boundary used by `next`.
+ */
+export function useCloseTransitionIntent(
+	current: BaseDescriptor,
+	animations: AnimationStoreMap,
+	system: SystemStoreMap,
+) {
+	const routeKey = current.route.key;
+	const { flags } = useStackCoreContext();
+	const { isBranchScreen, branchNavigatorKey } = useDescriptorDerivations();
+	const isNativeStack = flags.STACK_TYPE === StackType.NATIVE;
+	const enableHighRefreshRate =
+		current.options.experimental_enableHighRefreshRate ?? false;
+	const frameCallback = useFrameCallback(() => {}, false);
+
+	const activateHighRefreshRate = useStableCallback(() => {
+		if (enableHighRefreshRate) {
+			frameCallback.setActive(true);
+		}
+	});
+
+	const deactivateHighRefreshRate = useStableCallback(() => {
+		if (enableHighRefreshRate) {
+			frameCallback.setActive(false);
+		}
+	});
+
+	const resetStores = useStableCallback(() => {
+		resetStoresForScreen(routeKey, isBranchScreen, branchNavigatorKey);
+	});
+
+	const closeParams: CloseHookParams = {
+		current,
+		animations,
+		system,
+		resetStores,
+		activateHighRefreshRate,
+		deactivateHighRefreshRate,
+	};
+
+	if (isNativeStack) {
+		return useNativeStackClose(closeParams);
+	}
+
+	return useManagedClose(closeParams);
+}
